Combine project_dom_cache abd todos_dom_cache into one file, separate file for event_listeners
